moved monitoring code to another module
