+++ src/gridengineapp/__init__.py
@@ -5,7 +5,8 @@ from gridengineapp.graph_choice import execution_ordered
 from gridengineapp.identifier import IntegerIdentifier, StringIdentifier
 from gridengineapp.job import Job
 from gridengineapp.main import entry
-from gridengineapp.status import qstat, qstat_short, check_complete
+from gridengineapp.status import qstat, qstat_short
+from gridengineapp.monitor import check_complete
 from gridengineapp.submit import qsub, qsub_template
 from gridengineapp.gridparser import GridParser, ArgumentError
 
@@ -14,7 +15,6 @@ __all__ = [
     "qdel", "qstat", "qstat_short", "qsub", "qsub_template",
     "GridParser",
     "ArgumentError",
-    "check_complete",
     "FileEntity",
     "PandasFile",
     "ShelfFile",
